+++ ContentPatcher/docs/release-notes.md
@@ -11,6 +11,7 @@ When releasing a format change, don't forget to update the smapi.io/json schema!
 -->
 ## Upcoming release
 * Fixed patches which use the `i18n` token not always updated on language change.
+* Fixed `HasMod` condition not enabling mod-provided tokens within the same `When` block.
 
 ## 1.23.4
 Released 04 September 2021 for SMAPI 3.12.6 or later.